Don't allow rescheduling, use closed channel as message to stop.
/cc @armon - Can you just verify this is still correct? This is a lot
cleaner in terms of logic for stopping the tickers. Instead of sending
N stop messages on a channel, we just use a closed channel to signal
that it is over.

I modified triggerFunc to receive the stopCh as a parameter since there
is a subtle data race otherwise: deschedule and schedule() might be
called before the loop spins back around to read "m.stopTick" and
you might get the WRONG channel on that read. Pushing it in as a param
ensures the correct channel is always used.

Finally, I made schedule/deschedule idempotent.