Add EventDelegate and remote JoinCh and LeaveCh
The channels do not provide a logical ordering of join/leaves, and it
means a client could potentially get a leave before a join. To avoid
this, we instead use an EventDelegate which has NotifyJoin() and
NotifyLeave(). As a shim, we provide the ChannelEventDelegate which
allows events to still be received over a channel.