reduce NACK influence
Lifeguard introduced Self-Awareness, with an awareness delta that is in
part based on the difference between the expected and actual NACKs
received. A multiplier of 2 was used on this value. Benchmarking
indicates this may start to become counterproductive at very high
numbers of concurrent anomalies, so lets remove the multiplier.