reduce raciness by copying nodes in kRandomNodes
While kRandomNodes is usually called while the node list mutex is
locked, the resulting *node pointers are frequently dereferenced
without the mutex being locked. Some things, like AliveNode(),
modify the state of an existing node in a node map, which means
that access to that node's contents elsewhere is a race.
One solution is to ensure that we're copying these states out.
(The other would be for AliveNode and the like to replace the
object in the node map with a new one.)