introduce new state: stateLeft (#213)
This PR introduces another state: stateLeft. This is an attempt to help fix hashicorp/consul#6897.

When memberlist processes a dead message, it will now check where it is coming from and if the reporting node is equal to the reported node, we know this node wants to leave. In this case we don't have to prevent new nodes from using the same ip.

If a node reports another node, the behavior stays the same.