Add public key encryption
This adds the ability for nodes to encrypt traffic between them using
curve 25519 keys. Address is extended to allow a pubilc key to be
specified along with the nodename, ip, and port. This scheme reuses the
AES encryption bits, using curve 25519 Diffie-Hellman to simply compute
the key to encrypt the payloads with.

Additionally, if a secret access key is provided, it use mixed into the
encryption key. This allows the secret key to be used as a cluster
access token which is not exposed as encryption material and thusly does
not need to be rotated the same as it otherwise would be.