tests: Use net.JoinHostPort()
While starting to look at how memberlist is used by a project, I was
browsing around and happened to notice these lines that manually
joined the bind address and port.  While it's not actually causing a
problem with these tests, it's good practice to always use
net.JoinHostPort() to do this, which will ensure that IPv6 addresses
are handled correctly.