Fix data races in gossip()
The `gossip` function obtains a read lock for the nodes map, but
releases it once it gets the random sample of nodes. Since kNodes is
a slice of pointers to nodeState, reading any fields on nodeState could
race with any other goroutine that writes to the fields in nodeState.

The go -race detector found a few places that were correctly using the
nodeLock, but were racing because `gossip` was reading a pointer value.
See hashicorp/memberlist/issues/134.

There were two other callers to `kRandomNodes` that likely have the same
problem because they released the read lock in the same way.

This commit fixes the problem by changing `kRandomNodes` to return a
slice of `Node` instead of `*nodeState`. This allows the callers to read
fields of the Node without having to hold the read lock, and without
racing with any writers.

Once side effect of this change is that `kRandomNodes` can no longer
compare Nodes using pointer address. It was changed to compare Nodes
using the Node.Name, which as I understand it, should be unique. The
name is used as a key in the nodeMap.