Fix a data race in EventDelegate
EventDelegate is passed a pointer to a node that can be modified by other goroutines. This
causes a data race because the EventDelegate could be implemented and passed in by the caller.

This commit fixes the problem in tests, but the race still exists for all other callers.

It seems a more involved fix will be necessary, but I am fixing it this way for now to
address other data races.

The test had to be updated because the pointers no longer match.