Fix three data races in tests
Trying to use a logger that sends to t.Log is a problem when goroutines can outlive the
main test goroutine. It can cause the test to panic if a write happens after the test
exits.

In general there is no reason to send output to t.Log because 'go test' writes output to
stdout. Fix the data race by removing the logger in the integration test. Output will now
go to stdout.

Also fixes access to a field that should be read atomicly.

Finally the third test requires a lock to access the nodes.