Fix flaky tests
By waiting for the condition to be true, instead of sleeping a fixed amount. Occasionally these tests
would fail when it took too long to run.

Also remove the use of t.Log from a logger. This would cause tests to panic because the goroutines
using t.Log can outline the main test function.