Upgrade to go-msgpack/v2 2.1.1
This adds the option to specify what version of the `time.Time` encoding
format is desired. The default option is `false`, which preserves
compatibility with the current dependency of
`hashicorp/go-msgpack 0.5.3` in `go.mod`, but users of this library will
probably want to override the the setting.

While we're here, test against the latest versions of Go and
upgrade the `go.mod` file.

I tested that this appears to work with Nomad.