fix skip label check
Memberlist allows having label and to check them on receiving a message
to ensure rings don't collide unexpectedly. The option to skip that
checks currently doesn't work and will fail the message unless the
labels are exactly the same.

This PR change that behaviour to ensure that the option to skip the
label check does that.

Signed-off-by: Loic Reyreaud <loic@weaviate.io>