Export node State to NotifyLeave callback
The node State was exported to the application in commit
1c2989eaa138f8ce55e7d78384dfdc292c04dcb7, but only the MergeDelegate is
passed Node values with the State field initialized. The actual state of
each node as tracked by the local memberlist is hidden from the
application in the non-exported nodeState struct which embeds Node.
Remove the State field from nodeState so that the State field on the
embedded Node is instead used to track the node state. This has the
effect of making the actual state of the node (Left vs. Dead) available
to NotifyLeave callbacks as a pointer to the nodeState's embedded Node
struct is passed to EventDelegate callbacks.